Re-implemented the web crawler logic using Niocchi as the crawler.
git-svn-id: http://ikube.googlecode.com/svn/trunk@1912 89a0fb20-889e-aefa-90f6-7c296b20d5cf

+++ code/com/src/main/java/ikube/toolkit/FileUtilities.java
@@ -445,7 +445,7 @@ public final class FileUtilities {
      * @param file the file to read the contents from
      * @return the file contents in a byte array output stream
      */
-    public static ByteArrayOutputStream getContents(final File file, final int maxReadLength) {
+    public static ByteArrayOutputStream getContents(final File file, final long maxReadLength) {
         InputStream inputStream = null;
         try {
             inputStream = new FileInputStream(file);
